Refactor BfacDialog into LoginPanel, SubmissionPanel, and BfacBackendSync
Extract login UI, submission form, and background API sync into separate
components so BfacDialog only handles navigation and upload lifecycle.
